Module aws_sdk_greengrassv2::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

request_already_in_progress_exception

See RequestAlreadyInProgressException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You don't have permission to perform the action.

BatchAssociateClientDeviceWithCoreDeviceError

BatchDisassociateClientDeviceFromCoreDeviceError

CancelDeploymentError

ConflictException

Your request has conflicting operations. This can occur if you're trying to perform more

CreateComponentVersionError

CreateDeploymentError

DeleteComponentError

DeleteCoreDeviceError

DescribeComponentError

GetComponentError

GetComponentVersionArtifactError

GetCoreDeviceError

GetDeploymentError

InternalServerException

IoT Greengrass can't process your request right now. Try again later.

ListClientDevicesAssociatedWithCoreDeviceError

ListComponentVersionsError

ListComponentsError

ListCoreDevicesError

ListDeploymentsError

ListEffectiveDeploymentsError

ListInstalledComponentsError

ListTagsForResourceError

RequestAlreadyInProgressException

The request is already in progress. This exception occurs when you use a client token for

ResolveComponentCandidatesError

ResourceNotFoundException

The requested resource can't be found.

ServiceQuotaExceededException

Your request exceeds a service quota. For example, you might have the maximum number of

TagResourceError

ThrottlingException

Your request exceeded a request rate quota. For example, you might have exceeded the

UntagResourceError

ValidationException

The request isn't valid. This can occur if your request contains malformed JSON or

Enums

BatchAssociateClientDeviceWithCoreDeviceErrorKind

BatchDisassociateClientDeviceFromCoreDeviceErrorKind

CancelDeploymentErrorKind

CreateComponentVersionErrorKind

CreateDeploymentErrorKind

DeleteComponentErrorKind

DeleteCoreDeviceErrorKind

DescribeComponentErrorKind

GetComponentErrorKind

GetComponentVersionArtifactErrorKind

GetCoreDeviceErrorKind

GetDeploymentErrorKind

ListClientDevicesAssociatedWithCoreDeviceErrorKind

ListComponentVersionsErrorKind

ListComponentsErrorKind

ListCoreDevicesErrorKind

ListDeploymentsErrorKind

ListEffectiveDeploymentsErrorKind

ListInstalledComponentsErrorKind

ListTagsForResourceErrorKind

ResolveComponentCandidatesErrorKind

TagResourceErrorKind

UntagResourceErrorKind