Module aws_sdk_greengrassv2::model

Modules

associate_client_device_with_core_device_entry

See AssociateClientDeviceWithCoreDeviceEntry

associate_client_device_with_core_device_error_entry

See AssociateClientDeviceWithCoreDeviceErrorEntry

associated_client_device

See AssociatedClientDevice

cloud_component_status

See CloudComponentStatus

component

See Component

component_candidate

See ComponentCandidate

component_configuration_update

See ComponentConfigurationUpdate

component_dependency_requirement

See ComponentDependencyRequirement

component_deployment_specification

See ComponentDeploymentSpecification

component_latest_version

See ComponentLatestVersion

component_platform

See ComponentPlatform

component_run_with

See ComponentRunWith

component_version_list_item

See ComponentVersionListItem

core_device

See CoreDevice

deployment

See Deployment

deployment_component_update_policy

See DeploymentComponentUpdatePolicy

deployment_configuration_validation_policy

See DeploymentConfigurationValidationPolicy

deployment_io_t_job_configuration

See DeploymentIoTJobConfiguration

deployment_policies

See DeploymentPolicies

disassociate_client_device_from_core_device_entry

See DisassociateClientDeviceFromCoreDeviceEntry

disassociate_client_device_from_core_device_error_entry

See DisassociateClientDeviceFromCoreDeviceErrorEntry

effective_deployment

See EffectiveDeployment

installed_component

See InstalledComponent

io_t_job_abort_config

See IoTJobAbortConfig

io_t_job_abort_criteria

See IoTJobAbortCriteria

io_t_job_executions_rollout_config

See IoTJobExecutionsRolloutConfig

io_t_job_exponential_rollout_rate

See IoTJobExponentialRolloutRate

io_t_job_rate_increase_criteria

See IoTJobRateIncreaseCriteria

io_t_job_timeout_config

See IoTJobTimeoutConfig

lambda_container_params

See LambdaContainerParams

lambda_device_mount

See LambdaDeviceMount

lambda_event_source

See LambdaEventSource

lambda_execution_parameters

See LambdaExecutionParameters

lambda_function_recipe_source

See LambdaFunctionRecipeSource

lambda_linux_process_params

See LambdaLinuxProcessParams

lambda_volume_mount

See LambdaVolumeMount

resolved_component_version

See ResolvedComponentVersion

system_resource_limits

See SystemResourceLimits

validation_exception_field

See ValidationExceptionField

Structs

AssociateClientDeviceWithCoreDeviceEntry

Contains a request to associate a client device with a core device. The BatchAssociateClientDeviceWithCoreDevice operation consumes a list of these

AssociateClientDeviceWithCoreDeviceErrorEntry

Contains an error that occurs from a request to associate a client device with a core

AssociatedClientDevice

Contains information about a client device that is associated to a core device for cloud

CloudComponentStatus

Contains the status of a component in the IoT Greengrass service.

Component

Contains information about a component.

ComponentCandidate

Contains information about a component that is a candidate to deploy to a Greengrass core

ComponentConfigurationUpdate

Contains information about a deployment's update to a component's configuration on

ComponentDependencyRequirement

Contains information about a component dependency for a Lambda function component.

ComponentDeploymentSpecification

Contains information about a component to deploy.

ComponentLatestVersion

Contains information about the latest version of a component.

ComponentPlatform

Contains information about a platform that a component supports.

ComponentRunWith

Contains information system user and group that the IoT Greengrass Core software uses to run component

ComponentVersionListItem

Contains information about a component version in a list.

CoreDevice

Contains information about a Greengrass core device, which is an IoT thing that runs the IoT Greengrass

Deployment

Contains information about a deployment.

DeploymentComponentUpdatePolicy

Contains information about a deployment's policy that defines when components are safe to

DeploymentConfigurationValidationPolicy

Contains information about how long a component on a core device can validate its

DeploymentIoTJobConfiguration

Contains information about an IoT job configuration.

DeploymentPolicies

Contains information about policies that define how a deployment updates components and

DisassociateClientDeviceFromCoreDeviceEntry

Contains a request to disassociate a client device from a core device. The BatchDisassociateClientDeviceWithCoreDevice operation consumes a list of these

DisassociateClientDeviceFromCoreDeviceErrorEntry

Contains an error that occurs from a request to disassociate a client device from a core

EffectiveDeployment

Contains information about a deployment job that IoT Greengrass sends to a Greengrass core device.

InstalledComponent

Contains information about a component on a Greengrass core device.

IoTJobAbortConfig

Contains a list of criteria that define when and how to cancel a configuration

IoTJobAbortCriteria

Contains criteria that define when and how to cancel a job.

IoTJobExecutionsRolloutConfig

Contains information about the rollout configuration for a job. This configuration defines

IoTJobExponentialRolloutRate

Contains information about an exponential rollout rate for a configuration deployment

IoTJobRateIncreaseCriteria

Contains information about criteria to meet before a job increases its rollout rate.

IoTJobTimeoutConfig

Contains information about the timeout configuration for a job.

LambdaContainerParams

Contains information about a container in which Lambda functions run on Greengrass core

LambdaDeviceMount

Contains information about a device that Linux processes in a container can access.

LambdaEventSource

Contains information about an event source for an Lambda function. The event source

LambdaExecutionParameters

Contains parameters for a Lambda function that runs on IoT Greengrass.

LambdaFunctionRecipeSource

Contains information about an Lambda function to import to create a component.

LambdaLinuxProcessParams

Contains parameters for a Linux process that contains an Lambda function.

LambdaVolumeMount

Contains information about a volume that Linux processes in a container can access. When

ResolvedComponentVersion

Contains information about a component version that is compatible to run on a Greengrass core

SystemResourceLimits

Contains information about system resource limits that the IoT Greengrass Core software applies to a

ValidationExceptionField

Contains information about a validation exception field.

Enums

CloudComponentState

ComponentDependencyType

ComponentVisibilityScope

CoreDeviceStatus

DeploymentComponentUpdatePolicyAction

DeploymentFailureHandlingPolicy

DeploymentHistoryFilter

DeploymentStatus

EffectiveDeploymentExecutionStatus

InstalledComponentLifecycleState

IoTJobAbortAction

IoTJobExecutionFailureType

LambdaEventSourceType

LambdaFilesystemPermission

LambdaInputPayloadEncodingType

LambdaIsolationMode

RecipeOutputFormat

ValidationExceptionReason