Module aws_sdk_greengrassv2::operation

Structs

BatchAssociateClientDeviceWithCoreDevice

Associate a list of client devices with a core device. Use this API operation to specify

BatchDisassociateClientDeviceFromCoreDevice

Disassociate a list of client devices from a core device. After you disassociate a client

CancelDeployment

Cancels a deployment. This operation cancels the deployment for devices that haven't yet

CreateComponentVersion

Creates a component. Components are software that run on Greengrass core devices. After you

CreateDeployment

Creates a continuous deployment for a target, which is a Greengrass core device or group of core

DeleteComponent

Deletes a version of a component from IoT Greengrass.

DeleteCoreDevice

Deletes a Greengrass core device, which is an IoT thing. This operation removes the core

DescribeComponent

Retrieves metadata for a version of a component.

GetComponent

Gets the recipe for a version of a component. Core devices can call this operation to

GetComponentVersionArtifact

Gets the pre-signed URL to download a public component artifact. Core devices call this

GetCoreDevice

Retrieves metadata for a Greengrass core device.

GetDeployment

Gets a deployment. Deployments define the components that run on Greengrass core devices.

ListClientDevicesAssociatedWithCoreDevice

Retrieves a paginated list of client devices that are associated with a core

ListComponentVersions

Retrieves a paginated list of all versions for a component. Greater versions are listed first.

ListComponents

Retrieves a paginated list of component summaries. This list includes components that you

ListCoreDevices

Retrieves a paginated list of Greengrass core devices.

ListDeployments

Retrieves a paginated list of deployments.

ListEffectiveDeployments

Retrieves a paginated list of deployment jobs that IoT Greengrass sends to Greengrass core

ListInstalledComponents

Retrieves a paginated list of the components that a Greengrass core device runs.

ListTagsForResource

Retrieves the list of tags for an IoT Greengrass resource.

ResolveComponentCandidates

Retrieves a list of components that meet the component, version, and platform requirements

TagResource

Adds tags to an IoT Greengrass resource. If a tag already exists for the resource, this operation

UntagResource

Removes a tag from an IoT Greengrass resource.